pyproject.toml
[tool.poetry]
name = "chrome-gpt"
version = "0.1.0"
description = "An AutoGPT agent that uses Chrome interactively on your desktop"
authors = ["Richard He <he.yucheng@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chrome_gpt"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = "^0.0.149"
selenium = "^4.9.0"
pytest = "^7.3.1"
mypy = "^1.2.0"
black = "^23.3.0"
ruff = "^0.0.263"
faiss-cpu = "^1.7.4"
click = "^8.1.3"
openai = "^0.27.4"
tiktoken = "^0.3.3"

[tool.ruff]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]
exclude = [
  "tests/integration_tests/examples/non-utf8-encoding.py",
]

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
exclude = ["notebooks"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
